config param improved

+++ motorised_hokuyo/src/sus_hokuyo.cpp
@@ -15,22 +15,17 @@ using namespace std;
 void topicHokuyo(const sensor_msgs::LaserScan::ConstPtr& scan_hokuyo) 
 {
 	
-	ROS_INFO("TOPIC NUEVO: %d", scan_hokuyo->ranges.size());
+	ROS_INFO("TOPIC NUEVO: %d", (int)scan_hokuyo->ranges.size());
    //scan.ranges = scan_hokuyo->ranges;
 	//ranges.push_back(scan_hokuyo->ranges);
    
 }
 
-
-
-
-
-
  int main(int argc, char **argv)
  {
    ros::init(argc, argv, "sus_hokuyo");
-	ros::NodeHandle m;
-	ros::Subscriber sus_hokuyo = m.subscribe("topic_hokuyo", 1, topicHokuyo); 
+	 ros::NodeHandle m;
+	 ros::Subscriber sus_hokuyo = m.subscribe("topic_hokuyo", 1, topicHokuyo); 
 
-	ros::spin();
+	 ros::spin();
  }
